Factorize ExpressionVariableParentFinder and ExpressionVariableTypeFinder.
