--- GDevelop.js/types/gdexpressionvariabletypefinder.js
@@ -1,7 +1,0 @@
-// Automatically generated by GDevelop.js/scripts/generate-types.js
-declare class gdExpressionVariableTypeFinder {
-  static getVariableType(platform: gdPlatform, projectScopedContainers: gdProjectScopedContainers, node: gdExpressionNode, objectName: string): Variable_Type;
-  static getArrayVariableType(platform: gdPlatform, projectScopedContainers: gdProjectScopedContainers, node: gdExpressionNode, objectName: string): Variable_Type;
-  delete(): void;
-  ptr: number;
-};
