+++ GDevelop.js/types/libgdevelop.js
@@ -66,7 +66,6 @@ declare class libGDevelop {
   PairStringVariable: Class<gdPairStringVariable>;
   Variable_Type: Class<Variable_Type>;
   VariableInstructionSwitcher: Class<gdVariableInstructionSwitcher>;
-  ExpressionVariableTypeFinder: Class<gdExpressionVariableTypeFinder>;
   Variable: Class<gdVariable>;
   VariablesContainer_SourceType: Class<VariablesContainer_SourceType>;
   VariablesContainer: Class<gdVariablesContainer>;
